requirements.txt
streamlit
pandas
pandas-datareader
google-generativeai
plotly
requests